update inits, move sprof to separate file
