+++ bgcArgoDMQC/core/__init__.py
@@ -1,1 +1,2 @@
-from __future__ import absolute_import
+from .core import *
+from .sprof import sprof
